python/raspberrypi-sensor-sound/requirements.txt
pygame
rpi.gpio
